Start to behaviour tree
I was trying to make/leanr behvaiour trees but realized I wont need that for this project since the enemy AI almost only needs to cycle through some animations and attacks.

I will keep everything I made in the project if I have too much time on my hands and want to continue learning behaviour trees.

Also made it so the "torch" state of the wand also shoots projectiles since we are gonna change the torches to chandeliers that you need to shoot to light up the place