Naming conventions, enemy ai, tick timer, sfx
changed some private variable names to have proper naming conventions. I haven't changed every variable yet.

Made an enemy AI script that handles the move selection based on weight of the move and it should handle IK smoothly.

made a tick timer for some functions that need to be called frequently but are heavy to put in the update. Using it for target finder now to turn on and off the targetting particle on objects

added some SFX for torches and fire projectile. also trying to make sfx at home