Animation mode culling fix and more
There was something wrong with the last animation push, wich caused older animation to cull the model. bc of this I couldnt work on any of the animation events since the animations/model need to be exported again.

added inputs for the barrier and the teleport.

Made the main scene into an art showcase since I couldn't implement the functionality because of the animation problem.